7.3.1
https://www.state-machine.com/qpc/history.html#qpc_7_3_1

+++ examples/posix-win32/defer/bsp.c
@@ -77,7 +77,7 @@ void QF_onClockTick(void) {
 
     key = QF_consoleGetKey();
     if (key != 0) { // any key pressed?
-        BSP_onKeyboardInput(key);
+        BSP_onKeyboardInput((uint8_t)key);
     }
 }
 //----------------------------------------------------------------------------
